+++ src/components/Home.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react'
-import { baseUrl } from '../shared/baseUrl'
-import { getUser, setUser, userFailed, userLoading } from '../redux/ActionCreators'
+import { userLoading, getUser, handleGetUser } from '../actions/userAction'
 import axios from 'axios'
 import { apiUrl} from '../shared/baseUrl'
 import { connect } from 'react-redux'
@@ -13,13 +12,7 @@ class Home extends Component {
   componentDidMount () {
     const {dispatch} = this.props;
     dispatch(userLoading())
-    const users = axios
-      .get(`${apiUrl}users`, )
-      .then(response => response.data)
-      .then(data => dispatch(getUser(data)))
-      .catch(error => {
-        console.log(error)
-      })
+    dispatch(handleGetUser())
   }
 
   render () {
