+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import userReducer from './userReducer'
+
+export default combineReducers({
+  user: userReducer,
+})
