--- src/redux/ActionTypes.js
@@ -1,5 +1,0 @@
-export const USER_LOADING = 'USER_LOADING'
-export const USER_LOGIN = 'USER_LOGIN'
-export const USER_FAILED = 'USER_FAILED'
-export const USER_LOADED = 'USER_LOADED'
-
